+++ config/services/message_generator.php
@@ -4,16 +4,15 @@ declare(strict_types=1);
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
-use Umanit\TwoFactorSms\Texter\NotifierAuthCodeTexter;
+use Umanit\TwoFactorSms\Texter\SmsMessageGenerator;
 
 return static function (ContainerConfigurator $container): void {
     $services = $container->services();
 
     $services
-        ->set('umanit_two_factor_sms.texter.auth_code_texter', NotifierAuthCodeTexter::class)
+        ->set('umanit_two_factor_sms.texter.sms_message_generator', SmsMessageGenerator::class)
         ->args([
-            service('texter')->ignoreOnInvalid(),
-            abstract_arg('message generator')
+            abstract_arg('message generator'),
         ])
     ;
 };