+++ src/Texter/NotifierAuthCodeTexter.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Umanit\TwoFactorSms\Texter;
 
-use Symfony\Component\Notifier\Message\SmsMessage;
 use Symfony\Component\Notifier\TexterInterface;
 use Umanit\TwoFactorSms\Model\Sms\TwoFactorSmsInterface;
 
@@ -12,7 +11,7 @@ final readonly class NotifierAuthCodeTexter implements AuthCodeTexterInterface
 {
     public function __construct(
         private TexterInterface $texter,
-        private string $message,
+        private SmsMessageGeneratorInterface $messageGenerator,
     ) {
     }
 
@@ -23,16 +22,6 @@ final readonly class NotifierAuthCodeTexter implements AuthCodeTexterInterface
             return;
         }
 
-        $sms = new SmsMessage(
-            $user->getSmsAuthRecipient(),
-            $this->getMessage($authCode),
-        );
-
-        $this->texter->send($sms);
-    }
-
-    public function getMessage(string $authCode): string
-    {
-        return str_replace('[[auth_code]]', $authCode, $this->message);
+        $this->texter->send($this->messageGenerator->generateMessage($user));
     }
 }