+++ src/Texter/SmsMessageGeneratorInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Umanit\TwoFactorSms\Texter;
+
+use Symfony\Component\Notifier\Message\SmsMessage;
+use Umanit\TwoFactorSms\Model\Sms\TwoFactorSmsInterface;
+
+interface SmsMessageGeneratorInterface
+{
+    public function generateMessage(TwoFactorSmsInterface $user): SmsMessage;
+}